Create random player
To create a random play the code was redesigned to make the make_move function linked to the player class. The AlignFive, GameBoard and GameWindow code was aligned with this. 
GameWindow is now a property of the GameBoard rather than being called in AlignFive.play_game.
Move dataclass was added to provide access to a Position and a player_number.