Remove board_size attribute and replace with use of shape function
Setting the board_size attribute caused issues in the code. This has been replaced by the use of .shape on the board attribute (which is a numpy array). This also means the code is more generic and would work on a rectangular matrix as well. 
The tests were updated to reflect this change.